requirements.txt
numpy
tensorflow
librosa
pyaudio
python-rtmidi
